Remove cl building warnings

+++ src/sm4_ccm.c
@@ -74,7 +74,7 @@ int sm4_ccm_encrypt(const SM4_KEY *sm4_key, const uint8_t *iv, size_t ivlen,
 	}
 
 	inlen_size = 15 - ivlen;
-	if (inlen_size < 8 && inlen >= (1 << (inlen_size * 8))) {
+	if (inlen_size < 8 && inlen >= ((size_t)1 << (inlen_size * 8))) {
 		error_print();
 		return -1;
 	}
